Quater5/02-prompt-engineering/pyproject.toml
[tool.poetry]
name = "02-prompt-engineering"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-google-genai = "^2.0.9"
python-multipart = "^0.0.20"
langchain = "^0.3.15"
langchain-core = "^0.3.31"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
